abcMIDI-2013-09-08
September 04 2013

Midicopy: new feature. Introduced the run time option -tempo n (beats/min). When
this option is included all tempo messages in the midi file are
replaced with the specified value.

Implementation: in metaevent() when a tempo message is encountered,
if newtempo is greater than zero, write the newtempo value instead
of the value in the source midifile. The value of newtempo is set
by the -tempo run time option.

September 06 2013

Midicopy: new feature. Introduced the run time option -speed f (multiplier).
When this option is included all tempo messages in the midi file are
are multiplied by the factor f where f is a floating point number between
0.05 and 10.0. This speeds up or slows the music by this factor.

September 07 2013

Midicopy: new feature. Introduced the run time option -drumfocus n m
where n is the pitch code of the selected drum line and m is the velocity
value to be applied to all other drum lines. The feature, highlights
the selected drum line by attenuating all other drum lines by playing
them at level m (m varies from 0 to 127).

Implementation: the velocities of the other drum lines are set to
m in function chanmessage() prior to sending a 'note on' message.

September 08 2013

Midicopy: fixed function tick_to_seconds so it returns the correct
time when the tempo is changed by -tempo or -speed.